added Sushi interface
