+++ frontend/src/interfaces/Sushi.ts
@@ -1,0 +1,7 @@
+export interface Sushi {
+    item_name: string,
+    item_price: number,
+    item_kcal: number,
+    item_category: string,
+    is_alcohol: boolean,
+}
